Adding change request headings to clause_4 as a starting point
To provide a starting point for people to summarise what else we
need to reduce barriers and increase uptake of semantic and graph
technologies, a template for adding and summarising the existing
change requests in standards tracker was added. The summary lines
were used for this. The detail has not been entered.