Fix dangling references copied from reference 3 in clause 2 UCs
There were a number of dangling references copied from reference 3
that resulted in confusion. These have since been resolved by
substituting the organisations they refer to, or removing them in
order to draw attention to the original reference. These use cases
are more likely to be removed in time, or new and true references
added in their place.

Fixes #21