Removed Web3D from the list of related groups due to tenuous link
At this stage at least, we cannot see any reason to include the
Web3D group as a related activity. We can watch the group to
determine whether we should engage with them in future, but it's
not immediately apparent what their relevance to GeoSPARQL is
based on the knowledge we have right now.