Added a number of relationships with current OGC standards
As stated, added DGGS and CRS as relationships with current OGC
standards, as they have been specifically state in change requests
and are likely to be explored at least.

Also updated Time Ontology to include OGC related work. Removed
discussion of non-OGC work, as this is not the purpose of this
section.